backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14,<4
django-cors-headers>=4.3,<5
python-dotenv>=1.0,<2
Pillow>=10.0,<11
